classification-service/requirements.txt
kafka-python
boto3
pandas
ultralytics
pillow
numpy
fastapi
aiokafka